flaskFinalProject/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
PySocks==1.7.1
requests==2.31.0
requests-file==2.0.0
SQLAlchemy==2.0.29
tldextract==5.1.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.2
